Fix caching, cache all params
The earlier caching implementation had a bug that sometimes caused
params not to get updated. Add a struct for params to track both the
results of a calculation using the passed value and Csound's current
knowledge of the param. Use this for all float-based params.

In addition to fixing the bug, this improves performance by 30%.